put focusNote and moveNoteToFront into utils
