+++ src/components/note.js
@@ -31,17 +31,3 @@ export function addNote(x, y) {
 export function removeNote(note) {
   document.querySelector(".note-gallery").removeChild(note);
 }
-
-export function moveNoteToFront(note) {
-  const notes = document.querySelectorAll(".note");
-  const highestZIndex = Array.from(notes).reduce((maxZIndex, note) => {
-    const zIndex = parseInt(window.getComputedStyle(note).zIndex, 10);
-    return isNaN(zIndex) ? maxZIndex : Math.max(maxZIndex, zIndex);
-  }, 0);
-
-  note.style.zIndex = highestZIndex + 1;
-}
-
-export function focusNote(note) {
-  note.querySelector(".note-text").focus();
-}
