+++ src/utils.js
@@ -1,9 +1,21 @@
-import { moveNoteToFront, focusNote } from "./components/note";
-
 export function focusElement(elmnt) {
   const isNote = elmnt.classList.contains("note");
   if (isNote) elmnt.querySelector(".note-text").focus();
   else elmnt.focus();
+}
+
+export function moveNoteToFront(note) {
+  const notes = document.querySelectorAll(".note");
+  const highestZIndex = Array.from(notes).reduce((maxZIndex, note) => {
+    const zIndex = parseInt(window.getComputedStyle(note).zIndex, 10);
+    return isNaN(zIndex) ? maxZIndex : Math.max(maxZIndex, zIndex);
+  }, 0);
+
+  note.style.zIndex = highestZIndex + 1;
+}
+
+export function focusNote(note) {
+  note.querySelector(".note-text").focus();
 }
 
 export function dragElement(elmnt) {
